Extract all of the numerical conversions into a separate module and leverage them where appropriate in AVM2
